backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
python-dotenv
pandas

